requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.1.3
marshmallow==3.21.3
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.2.0
PyMySQL==1.1.1
cryptography==42.0.8
